Backend/requirements.txt
fastapi==0.109.0
uvicorn==0.25.0
langchain==0.1.0
langchain-community==0.0.12
chromadb==0.4.22
sentence-transformers==2.2.2
langchain-google-genai==0.0.6
pypdf==3.17.4
cohere==4.10.0
python-dotenv==1.0.0
pdfplumber
